requirements.txt
openai
python-dotenv
yfinance
pillow
requests
typing